requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Google Gemini AI
google-genai>=1.0.0

# Supabase
supabase>=2.0.0

# HTTP Client
httpx>=0.27.0

# Scheduler
apscheduler>=3.10.0

# Environment
python-dotenv>=1.0.0

# Timezone
pytz>=2024.1

# Error Tracking
sentry-sdk[fastapi]>=2.0.0
